2D Added
We can now access multidimensional data. For now it only works for 2D
data. Need to figure out a way to auto gen for n-D data. But for now we
will be able to handle most cases. Which are 1D/2D int/float/double
data. I can confirm that data works when written and read with this
library. Need more testing from data generated from other sources.

+++ vecTest.cpp
@@ -10,7 +10,7 @@ using namespace H5;
 #define dim2 3
 int main()
 {
-   vector<vector<double> > v(dim1, vector<double>(dim2));
+   vector<vector<float> > v(dim1, vector<float>(dim2));
    for ( size_t i = 0; i < dim1; ++i )
       for ( size_t j = 0; j < dim2; ++j )
          v[i][j] = 3.;
@@ -61,7 +61,7 @@ int main()
    LoadH5 data;
    data.setFileName("dim.h5");
    data.setVarName("vecvecdouble");
-   vector<vector<double> > vv = data.getData();
+   vector<vector<float> > vv = data.getData();
    cout << "Data is read" << endl;
    for ( size_t i = 0; i < dim1; ++i )
    {
